Refactor: separate mint tokens action saga
