+++ src/modules/dashboard/sagas/actions.ts
@@ -4,9 +4,6 @@ import { ClientType, ROOT_DOMAIN_ID, ColonyVersion } from '@colony/colony-js';
 
 import { ContextModule, TEMP_getContext } from '~context/index';
 import {
-  TokenBalancesForDomainsDocument,
-  TokenBalancesForDomainsQuery,
-  TokenBalancesForDomainsQueryVariables,
   ColonyFromNameDocument,
   ColonyFromNameQuery,
   ColonyFromNameQueryVariables,
@@ -29,152 +26,6 @@ import {
   transactionAddParams,
 } from '../../core/actionCreators';
 import { updateColonyDisplayCache } from './utils';
-
-function* createMintTokensAction({
-  payload: {
-    colonyAddress,
-    colonyName,
-    nativeTokenAddress,
-    amount,
-    annotationMessage,
-  },
-  meta: { id: metaId, history },
-  meta,
-}: Action<ActionTypes.COLONY_ACTION_MINT_TOKENS>) {
-  let txChannel;
-  try {
-    const apolloClient = TEMP_getContext(ContextModule.ApolloClient);
-
-    if (!amount) {
-      throw new Error('Amount to mint not set for mintTokens transaction');
-    }
-
-    txChannel = yield call(getTxChannel, metaId);
-
-    // setup batch ids and channels
-    const batchKey = 'mintTokens';
-
-    const {
-      mintTokens,
-      claimColonyFunds,
-      annotateMintTokens,
-    } = yield createTransactionChannels(metaId, [
-      'mintTokens',
-      'claimColonyFunds',
-      'annotateMintTokens',
-    ]);
-
-    // create transactions
-    yield fork(createTransaction, mintTokens.id, {
-      context: ClientType.ColonyClient,
-      methodName: 'mintTokens',
-      identifier: colonyAddress,
-      params: [amount],
-      group: {
-        key: batchKey,
-        id: metaId,
-        index: 0,
-      },
-      ready: false,
-    });
-    yield fork(createTransaction, claimColonyFunds.id, {
-      context: ClientType.ColonyClient,
-      methodName: 'claimColonyFunds',
-      identifier: colonyAddress,
-      params: [nativeTokenAddress],
-      group: {
-        key: batchKey,
-        id: metaId,
-        index: 1,
-      },
-      ready: false,
-    });
-
-    if (annotationMessage) {
-      yield fork(createTransaction, annotateMintTokens.id, {
-        context: ClientType.ColonyClient,
-        methodName: 'annotateTransaction',
-        identifier: colonyAddress,
-        params: [],
-        group: {
-          key: batchKey,
-          id: metaId,
-          index: 1,
-        },
-        ready: false,
-      });
-    }
-
-    yield takeFrom(mintTokens.channel, ActionTypes.TRANSACTION_CREATED);
-    yield takeFrom(claimColonyFunds.channel, ActionTypes.TRANSACTION_CREATED);
-    if (annotationMessage) {
-      yield takeFrom(
-        annotateMintTokens.channel,
-        ActionTypes.TRANSACTION_CREATED,
-      );
-    }
-
-    yield put(transactionReady(mintTokens.id));
-
-    const {
-      payload: { hash: txHash },
-    } = yield takeFrom(
-      mintTokens.channel,
-      ActionTypes.TRANSACTION_HASH_RECEIVED,
-    );
-    yield takeFrom(mintTokens.channel, ActionTypes.TRANSACTION_SUCCEEDED);
-    yield put(transactionReady(claimColonyFunds.id));
-    yield takeFrom(claimColonyFunds.channel, ActionTypes.TRANSACTION_SUCCEEDED);
-
-    if (annotationMessage) {
-      yield put(transactionPending(annotateMintTokens.id));
-
-      let ipfsHash = null;
-      ipfsHash = yield call(
-        ipfsUpload,
-        JSON.stringify({
-          annotationMessage,
-        }),
-      );
-
-      yield put(
-        transactionAddParams(annotateMintTokens.id, [txHash, ipfsHash]),
-      );
-
-      yield put(transactionReady(annotateMintTokens.id));
-
-      yield takeFrom(
-        annotateMintTokens.channel,
-        ActionTypes.TRANSACTION_SUCCEEDED,
-      );
-    }
-
-    yield apolloClient.query<
-      TokenBalancesForDomainsQuery,
-      TokenBalancesForDomainsQueryVariables
-    >({
-      query: TokenBalancesForDomainsDocument,
-      variables: {
-        colonyAddress,
-        tokenAddresses: [nativeTokenAddress],
-      },
-      fetchPolicy: 'network-only',
-    });
-
-    yield put<AllActions>({
-      type: ActionTypes.COLONY_ACTION_MINT_TOKENS_SUCCESS,
-      meta,
-    });
-
-    if (colonyName) {
-      yield routeRedirect(`/colony/${colonyName}/tx/${txHash}`, history);
-    }
-  } catch (caughtError) {
-    putError(ActionTypes.COLONY_ACTION_MINT_TOKENS_ERROR, caughtError, meta);
-  } finally {
-    txChannel.close();
-  }
-}
 
 function* createVersionUpgradeAction({
   payload: { colonyAddress, colonyName, version, annotationMessage },
@@ -1026,10 +877,6 @@ function* colonyTokenUnlock({
 }
 
 export default function* tasksSagas() {
-  yield takeEvery(
-    ActionTypes.COLONY_ACTION_MINT_TOKENS,
-    createMintTokensAction,
-  );
   yield takeEvery(ActionTypes.COLONY_ACTION_DOMAIN_CREATE, createDomainAction);
   yield takeEvery(
     ActionTypes.COLONY_ACTION_VERSION_UPGRADE,
