+++ src/modules/dashboard/sagas/actions/index.ts
@@ -2,7 +2,12 @@ import { all, call } from 'redux-saga/effects';
 
 import paymentActionSaga from './payment';
 import moveFundsActionSaga from './moveFunds';
+import mintTokensActionSaga from './mintTokens';
 
 export default function* actionsSagas() {
-  yield all([call(paymentActionSaga), call(moveFundsActionSaga)]);
+  yield all([
+    call(paymentActionSaga),
+    call(moveFundsActionSaga),
+    call(mintTokensActionSaga),
+  ]);
 }
